настроил сборщик и разбил на модули
