+++ src/components/load-button.js
@@ -1,0 +1,5 @@
+export const getLoadButton = () => {
+  return `
+    <button class="load-more" type="button">load more</button>
+  `;
+};
