+++ src/components/util.js
@@ -1,0 +1,1 @@
+export const getMarkup = (dataList, generator) => dataList.map(generator).join(`\n`);
